requirements.txt
openai
mysqlclient==2.2.4
mysql-connector-python==8.4.0
pydantic==2.7.1
fastapi==0.111.0
bcrypt==4.1.3
python-jose
httpx==0.27.0
python-dotenv~=1.0.1
uvicorn==0.29.0
starlette==0.37.2
pytest==8.2.0
pytest-mock==3.14.0
pytest-asyncio==0.23.6
jose~=1.0.0
pillow~=10.3.0
requests==2.32.3
yfinance
alpha_vantage